Long distance display support
With GPSOSD:
Max total trip distance travelled of > 12.5 miles now possible
Max total distance from home of > 12.5 miles now possible
With other FC:
Max distance = 12.5 miles (FC limitation)
Max total trip distance travelled of > 12.5 miles now possible

+++ MW_OSD/Screen.ino
@@ -142,7 +142,7 @@ void displayTemperature(void)        // DEPRECATED RUSHDUINO SUPPORT
 void displayMode(void)
 {
   
-  int16_t dist;
+  int32_t dist;
   if(Settings[S_UNITSYSTEM])
     dist = GPS_distanceToHome * 3.2808;           // mt to feet
   else
@@ -1372,7 +1372,7 @@ void mapmode(void) {
   int16_t angle;
   int16_t targetpos;
   int16_t centerpos;
-  uint16_t maxdistance;
+  uint32_t maxdistance;
   uint8_t mapsymbolcenter;
   uint8_t mapsymboltarget;
   uint8_t mapsymbolrange;
